Move "minimum WP version" related utilities to dedicated `MinimumWPVersionTrait`
The "minimum WP version" related utilities are only used by a small set of sniffs, so are better placed in a dedicated trait.

This moves the `$minimum_supported_version` property and the `get_wp_version_from_cl()` method to a new `WordPressCS\WordPress\Helpers\MinimumWPVersionTrait` and starts using that trait in the relevant sniffs.

Related to 1465
